requirement.txt
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.1
django-heroku==0.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
envs==1.4
gunicorn==20.1.0
psycopg2==2.9.3
pycryptodome==3.14.1
PyJWT==2.4.0
pymongo==4.2.0
pypiwin32==223
python-decouple==3.6
pytz==2021.3
selenium==3.141.0
sqlparse==0.4.2
urllib3==1.26.7
uvicorn==0.18.2
virtualenv==20.14.1
whitenoise==5.3.0

